http://jira.codehaus.org/browse/MODFORJ-211 This issues brought a problem to the front: when calling a CRUD create service with a complex DTO, there is no guarantee that the DTO returned is consistent. The DTO boject graph is created before the objects are flushed by Hibernate, which occurs at the end of the transaction, after the DTO translation. The work-around is to throw the DTO graph away and fetch it again with the read service.
To enforce this, the create service, and its associated domain service and DAO method, do not return the created object, but just the id. That can be used right away to read the object, in case you want to do more stuff with it.

The update service returns void, for more or less the same reasons.